Main task (#13)
* semaphore mutex, madgwick

can read both the mpu6050 and bmp280 at the "same " time.
also implements the Madgwick filter for the mpu6050

* Update main.cpp

* NEO6M Intergration

added
functions to init and get values using the tinygpsplus library

* changed serial port

* cleanup n stuff

* semaphore mutex, madgwick

can read both the mpu6050 and bmp280 at the "same " time.
also implements the Madgwick filter for the mpu6050

* NEO6M Intergration

added
functions to init and get values using the tinygpsplus library

* changed serial port

* cleanup n stuff

* Made changes to debug

* Created Main task and added buzzer mutex

* changed flight counters

---------

Co-authored-by: Isaac L <isaac.lay25@gmail.com>
Co-authored-by: Isaac Lay <isaaclay.25@gmail.com>